rename Categorie -> MededelingCategorie and enable caching
