+++ lib/model/mededelingen/MededelingCategorieenModel.class.php
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * MededelingCategorieenModel.class.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ *
+ */
+class MededelingCategorieenModel extends CachedPersistenceModel {
+
+	const ORM = 'MededelingCategorie';
+	const DIR = 'mededelingen/';
+
+	protected static $instance;
+	/**
+	 * Store MededelingCategorie array as a whole in memcache
+	 * @var boolean
+	 */
+	protected $memcache_prefetch = true;
+
+	/**
+	 * @param $categorie
+	 * @return false|MededelingCategorie
+	 */
+	public static function get($categorie) {
+		return static::instance()->retrieveByPrimaryKey(array($categorie));
+	}
+
+	/**
+	 * @return MededelingCategorie[]
+	 */
+	public static function getAll() {
+		return static::instance()->prefetch();
+	}
+
+}
